requirements.txt
autopep8==1.6.0
django-extensions==3.1.*
django-filter==21.1
django-seed==0.3.*
Django==3.2.*
djangorestframework==3.12.*
djhtml==1.4.11
djoser==2.1.0
dr-scaffold==2.1.*
drf-yasg==1.20.*
ipdb
psycopg2-binary==2.9.*
python-decouple==3.5
pytz
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1